% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runAppGoKidsGo}
\alias{runAppGoKidsGo}
\title{Run the GoKidsGoWeb Shiny Application}
\usage{
runAppGoKidsGo()
}
\description{
This function starts the Shiny application contained within the GoKidsGoWeb package.
It assumes that the application is stored within the 'shinyApp' directory in the 'inst' folder,
under the name 'app.R', which includes both the UI and server components of the Shiny app.
}
